Rename sanitiser/_id.js to sanitiser/_ids.js
